Make the Ninject Modules use conventions. Delete the Factory class as it is no longer needed and remove the CalendarItemToSingleTimeEntryConverter and CalendarItemToTimeRangeEntryConverter as they caused problems with the convention-based binding (moved their functionality into the Calendar/CalendarEntryBuilder)
