+++ DDDEastAnglia/DataAccess/EntityFramework/SessionVoteModelQuery.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using DDDEastAnglia.Models;
+
+namespace DDDEastAnglia.DataAccess.EntityFramework
+{
+    public class SessionVoteModelQuery : ISessionVoteModelQuery
+    {
+        private readonly IVoteRepository _voteRepository;
+        private readonly IConferenceRepository _conferenceRepository;
+
+        public SessionVoteModelQuery(IVoteRepository voteRepository, IConferenceRepository conferenceRepository)
+        {
+            _voteRepository = voteRepository;
+            _conferenceRepository = conferenceRepository;
+        }
+
+        public SessionVoteModel Get(int sessionId, Guid cookieId)
+        {
+            return new SessionVoteModel
+                {
+                    CanVote = _conferenceRepository.ForSession(sessionId).CanVote(),
+                    HasBeenVotedForByUser = _voteRepository.HasVotedFor(sessionId, cookieId),
+                    SessionId = sessionId
+                };
+        }
+    }
+}
