+++ DDDEastAnglia/DataAccess/MessageBus/CommandProcessingModule.cs
@@ -1,4 +1,4 @@
-﻿using DDDEastAnglia.DataAccess.Handlers.Voting;
+﻿using Ninject.Extensions.Conventions;
 using Ninject.Modules;
 using Ninject.Web.Common;
 
@@ -8,9 +8,20 @@ namespace DDDEastAnglia.DataAccess.MessageBus
     {
         public override void Load()
         {
-            Kernel.Bind<IHandle>().To<DeleteVoteCommandHandler>().InRequestScope();
-            Kernel.Bind<IHandle>().To<RegisterVoteCommandHandler>().InRequestScope();
-            Kernel.Bind<IMessageBus>().To<SimpleMessageBus>().InRequestScope();
+            // Once we have something better for handling the unit of work then these could move to being in the
+            // singleton scope as the handler defines the transaction boundary but until then the request 
+            // should be defining the transaction boundary.
+            Kernel.Bind(from => from.FromThisAssembly()
+                                    .SelectAllClasses()
+                                    .InNamespaceOf<CommandProcessingModule>()
+                                    .BindDefaultInterfaces()
+                                    .Configure(binding => binding.InRequestScope()));
+            
+            Kernel.Bind(from => from.FromThisAssembly()
+                                    .SelectAllClasses()
+                                    .InheritedFrom<IHandle>()
+                                    .BindDefaultInterfaces()
+                                    .Configure(binding => binding.InRequestScope()));
         }
     }
 }
