+++ DDDEastAnglia/packages.config
@@ -38,6 +38,7 @@
   <package id="Modernizr" version="2.6.2" targetFramework="net40" />
   <package id="Newtonsoft.Json" version="4.5.11" targetFramework="net40" />
   <package id="Ninject" version="3.0.1.10" targetFramework="net40" />
+  <package id="Ninject.Extensions.Conventions" version="3.0.0.11" targetFramework="net40" />
   <package id="Ninject.MVC3" version="3.0.0.6" targetFramework="net40" />
   <package id="Ninject.Web.Common" version="3.0.0.7" targetFramework="net40" />
   <package id="Nuget.Core" version="1.6.2" targetFramework="net40" />
